remove obsolete file
